Added Discount class, restructured pricingRules and added test suite.
I removed the discount information from the Item class and placed it into it's own class called Discount. Consequently, this meant I could restructure the pricingRules structure I was using to store items and their discounts. Instead of just storing Item objects, it is now a HashMap where the keys are the Item objects and the values are the Discount objects.

Also created the test suite to test the UnidaysDiscountChallenge class.